:bug: fix: itemService bug fix
